pyproject.toml
[project]
name = "data-eng-template"
version = "0.1.0"
description = ""
authors = [{ name = "Eyal Perry", email = "e.perry.pa@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "polars (>=1.26.0,<2.0.0)",
    "delta-spark (>=3.3.0,<4.0.0)",
    "pyspark (>=3.5.5,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
]

[tool.poetry]
packages = [{ include = "data_eng", from = "src" }]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
testcontainers = "^4.9.2"
kafka-python = "^2.1.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
